Storing uploaded image links in Mongo
First change was done to resolve the issue of not being
able to save the image into YelpCamp folder in cloudinary.
We do this by simply adding params in index.js of cloudinary.

Second was to remove the image in the model for campground and
make it into an array with url and filename.

Third, note that we removed the joi validation for images because
it will cause an error for validating it. Therefore, we won't be
validating it for the moment.

Fourth, we extracted the URL and filename using req.file from multer
in controller of campground.js

Fifth, we finally extracted the filename and url and used it in
show.ejs to display the image for each of the campground that we
make.