Added a custom popup by clicking a camp on map
The custom popup is not as easy as it seems to access
information to popup a particular required text.
In this case we popup an anchor tag referencing the
actual campground.
Note1: We have to use virtual to access it and also we have
to use something called 'opt' so that the popup is present in
the properties of the resulting object.
Note2: A minor change was done to increase the number of
campgrounds to 300.